Dedupe the direct-L* test stubs; drop the duplicate-code disable
The .pylintrc disable was for direct_lstar's duplication with the E-L* resolver,
which the strip removed -- the only duplicate-code left was the _StubTable /
_StubOracle / _pst fixtures shared by test_export_dfa and test_learner_duck_type.
Extract them to tests/direct_lstar_stubs and drop the disable: pylint is 10.00/10
with duplicate-code enabled again.
