+++ tests/direct_lstar_stubs.py
@@ -1,0 +1,36 @@
+"""Shared stubs for the direct-L* learner unit tests: a minimal table, an
+accept-everything oracle, and a PST namespace pointed at them."""
+
+from types import SimpleNamespace
+
+from orthogonal_dfa.l_star.memoized_oracle import MemoizedOracle
+
+
+class StubTable:
+    prefixes = ()
+
+    def suffix(self, v):
+        return [v]
+
+
+class StubOracle:
+    def membership_queries(self, strings):
+        # Accept everything, so a decisive re-read picks the accept side.
+        return [1] * len(strings)
+
+
+def make_pst():
+    oracle = StubOracle()
+    return SimpleNamespace(
+        alphabet_size=2,
+        accept_thresh=0.7,
+        reject_thresh=0.3,
+        decision_boundary=0.5,
+        evidence_margin=0.0,
+        table=StubTable(),
+        oracle=oracle,
+        sift_cache=MemoizedOracle(oracle),
+        config=SimpleNamespace(
+            split_pval=0.001, min_signal_strength=0.3, suffix_family_size=2
+        ),
+    )
